--- bioimageio_colab/__init__.py
@@ -1,1 +1,0 @@
-from bioimageio_colab.sam_annotation_server import start_sam_annotation_server as start_server
